pyproject.toml
[project]
name = "excel-mcp-server"
version = "0.1.7"
description = "Excel MCP Server for manipulating Excel files"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.10.1",
    "fastmcp>=2.0.0,<3.0.0",
    "openpyxl>=3.1.5",
    "typer>=0.16.0",
    "anthropic>=0.74.1",
    "python-dotenv>=1.2.1",
    "openai>=2.8.1",
    "dotenv>=0.9.9",
    "simplejson>=3.18.0",
    "bs4>=0.0.2",
    "browser-use>=0.9.7",
    "browser-use-sdk>=2.0.5",
    "pydantic>=2.12.4",
    "asyncio>=4.0.0",
    "playwright>=1.56.0",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "flask-cors>=6.0.1",
    "flask-sock>=0.7.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.4",
]
[[project.authors]]
name = "luCa"
email = "carollucagasan@gmail.com"

[project.scripts]
excel-mcp-server = "excel_mcp.__main__:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/excel_mcp"]

[tool.hatch.build]
packages = ["src/excel_mcp"]
